requirements.txt
ipykernel==6.26.0
asn1crypto==1.5.1
asttokens==2.4.1
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.1
cloudpickle==2.0.0
colorama==0.4.6
comm==0.2.0
config==0.5.1
cryptography==41.0.5
debugpy==1.8.0
decorator==5.1.1
et-xmlfile==1.1.0
exceptiongroup==1.1.3
executing==2.0.1
filelock==3.13.1
idna==3.4
importlib-metadata==6.8.0
ipykernel==6.26.0
ipython==8.17.2
jedi==0.19.1
jupyter_client==8.6.0
jupyter_core==5.5.0
matplotlib-inline==0.1.6
nest-asyncio==1.5.8
numpy==1.26.1
openpyxl==3.1.2
oscrypto==1.3.0
packaging==23.2
pandas==2.1.2
parso==0.8.3
platformdirs==3.11.0
prompt-toolkit==3.0.39
psutil==5.9.6
pure-eval==0.2.2
pycparser==2.21
pycryptodomex==3.19.0
Pygments==2.16.1
PyJWT==2.8.0
pyOpenSSL==23.3.0
python-dateutil==2.8.2
pytz==2023.3.post1
pywin32==306
PyYAML==6.0.1
pyzmq==25.1.1
requests==2.31.0
scipy==1.11.3
six==1.16.0
snowflake-connector-python==3.3.1
snowflake-snowpark-python==1.9.0
sortedcontainers==2.4.0
stack-data==0.6.3
tomlkit==0.12.1
tornado==6.3.3
traitlets==5.13.0
typing_extensions==4.8.0
tzdata==2023.3
urllib3==1.26.18
wcwidth==0.2.9
zipp==3.17.0
pyspark==3.5.0